+++ ui/opt/drm4g_gridway/include/gw_rm_msg.h
@@ -276,14 +276,14 @@ typedef struct gw_msg_host_s
 
     int size_mem_mb;
     int free_mem_mb;
-    int size_disk_mb;
-    int free_disk_mb;
+    long size_disk_mb;
+    long free_disk_mb;
 
     char fork_name[GW_MSG_STRING_SHORT];
     char lrms_name[GW_MSG_STRING_SHORT];
     char lrms_type[GW_MSG_STRING_SHORT];
 
-	int  number_of_queues;
+    int  number_of_queues;
     char queue_name[GW_HOST_MAX_QUEUES][GW_MSG_STRING_SHORT];
     int  queue_nodecount[GW_HOST_MAX_QUEUES];
     int  queue_freenodecount[GW_HOST_MAX_QUEUES];
@@ -296,32 +296,32 @@ typedef struct gw_msg_host_s
     char queue_dispatchtype[GW_HOST_MAX_QUEUES][GW_MSG_STRING_SHORT];
     char queue_priority[GW_HOST_MAX_QUEUES][GW_MSG_STRING_SHORT];
 
-	int  number_of_int_vars;   
-   	char gen_var_int_name[GW_HOST_MAX_GENVARS][GW_MSG_STRING_SHORT];
-   	int  gen_var_int_value[GW_HOST_MAX_GENVARS];
-
-	int  number_of_str_vars;
-   	char gen_var_str_name[GW_HOST_MAX_GENVARS][GW_MSG_STRING_SHORT];
-   	char gen_var_str_value[GW_HOST_MAX_GENVARS][GW_MSG_STRING_LONG];   	
+    int  number_of_int_vars;   
+    char gen_var_int_name[GW_HOST_MAX_GENVARS][GW_MSG_STRING_SHORT];
+    int  gen_var_int_value[GW_HOST_MAX_GENVARS];
+ 
+    int  number_of_str_vars;
+    char gen_var_str_name[GW_HOST_MAX_GENVARS][GW_MSG_STRING_SHORT];
+    char gen_var_str_value[GW_HOST_MAX_GENVARS][GW_MSG_STRING_LONG];   	
     
 } gw_msg_host_t;
 
 
 typedef struct gw_msg_match_s 
 {
-	gw_msg_type_t	 msg_type;
-	gw_return_code_t rc;
+    gw_msg_type_t	 msg_type;
+    gw_return_code_t rc;
 	
-	int              host_id;
-	int              job_id;
-	int              fixed_priority;
+    int              host_id;
+    int              job_id;
+    int              fixed_priority;
 
-	gw_boolean_t     matched;
+    gw_boolean_t     matched;
 
     char hostname[GW_MSG_STRING_HOST];
-	int  running_jobs;
+    int  running_jobs;
 	
-	int  number_of_queues;
+    int  number_of_queues;
     char queue_name[GW_HOST_MAX_QUEUES][GW_MSG_STRING_SHORT];
     int  match[GW_HOST_MAX_QUEUES];
     int  rank [GW_HOST_MAX_QUEUES];
@@ -338,19 +338,19 @@ typedef struct gw_msg_user_s
     int user_id;
     char name[GW_MSG_STRING_SHORT];	
     char dn[GW_MSG_STRING_LONG];	
-    
-	int active_jobs;
-	int running_jobs;
+     
+    int active_jobs;
+    int running_jobs;
 	
-	time_t idle;
+    time_t idle;
 	
-	int  num_ems;
-	int  em_pid[GW_MAX_MADS];
-	char em_name[GW_HOST_MAX_QUEUES][GW_MSG_STRING_SHORT];
+    int  num_ems;
+    int  em_pid[GW_MAX_MADS];
+    char em_name[GW_HOST_MAX_QUEUES][GW_MSG_STRING_SHORT];
 	
-	int  num_tms;
-	int  tm_pid[GW_MAX_MADS];
-	char tm_name[GW_HOST_MAX_QUEUES][GW_MSG_STRING_SHORT];
+    int  num_tms;
+    int  tm_pid[GW_MAX_MADS];
+    char tm_name[GW_HOST_MAX_QUEUES][GW_MSG_STRING_SHORT];
     
 } gw_msg_user_t;
